@@ -1,7 +1,7 @@
 package com.sub6resources.utilities
 
 import android.graphics.Color
-import android.support.v7.widget.Toolbar
+import androidx.appcompat.widget.Toolbar
 import android.view.LayoutInflater
 import android.view.View
 import android.widget.TextView
